Add files via upload
The Data Science field has become a very hot field over the past few years. Many people are looking to go into the data science field, with the hope of making a lot of money. Surprisingly, many do not know the diversity in the data science field. In this project, we will explore the various data science fields, as well as the salaries of each field. We will also use that information to predict salaries of the data science fields.


The dataset used in this study is taken from kaggle, an online platform for data science. The dataset contains:

1. work_year: The year the salary was paid
2. experience_level: The level of experience in the job during the year
3. job_title: The title of the job worked in during the year
4. employment_type: The type of job employment for the role
5. employee_residence: The primary country of residence of the employee during the work year as an ISO 3166 country code
6. remote_ratio: The ratio of remote work done
7. company_location: The primary country of residence of the company during the work year
8. company_size: The median number of employees workin in the company.
9. salary_in_usd: The salary of the employee in USD
10. salary_currency: The currency of the salary
11. salary: The gross salary paid.